+++ blueprint/billing-base/domain/mappers/paymentLink.mappers.ts
@@ -1,6 +1,6 @@
 import { schemaValidator } from '@forklaunch/blueprint-core';
 import { requestMapper, responseMapper } from '@forklaunch/core/mappers';
-import { EntityManager, InferEntity, wrap } from '@mikro-orm/core';
+import { EntityManager } from '@mikro-orm/core';
 import { PaymentLink } from '../../persistence/entities/paymentLink.entity';
 import { CurrencyEnum } from '../enum/currency.enum';
 import { PaymentMethodEnum } from '../enum/paymentMethod.enum';
@@ -51,9 +51,9 @@ export const PaymentLinkMapper = responseMapper({
   ),
   entity: PaymentLink,
   mapperDefinition: {
-    toDto: async (entity: InferEntity<typeof PaymentLink>) => {
+    toDto: async (entity) => {
       return {
-        ...wrap(entity).toPOJO(),
+        ...entity,
         amount: Number(entity.amount)
       };
     }