+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @forklaunch/interfaces-billing
 
+## 1.0.10
+
+### Patch Changes
+
+- package version increase
+
+## 1.0.9
+
+### Patch Changes
+
+- shorter brand for compliance entities
+
 ## 1.0.8
 
 ### Patch Changes