+++ framework/bunrun/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/bunrun
 
+## 1.2.2
+
+### Patch Changes
+
+- tenant and rls configuration
+
+## 1.2.1
+
+### Patch Changes
+
+- fix compliance entity
+
+## 1.2.0
+
+### Minor Changes
+
+- Validator 25% performance uptick and cleaner Config Injector syntax
+
 ## 1.1.8
 
 ### Patch Changes